Move archive dataset function to models.py
When testing udata-front (?), we have some circular import errors.
Moving the archive logic to models prevent these.
